Forward to 1.5.0 (#10)
* Changelog preparation for future 1.4.2 release

* convert: Fix panic: heterogeneous tuple with null

Tuples with elements of different types can be converted to homogeneous
collections (sets or lists), so long as their elements are unifiable.
For example:

  list("a", "b")     // all elements have the same type
  list("a", 5)       // "a" and 5 can be unified to string
  list("a", 5, null) // null is a valid value for string

However, tuples with elements which are not unifiable cannot be
converted to homogeneous collections:

  list(["a"], "b")   // no common type for list(string) and string

This commit fixes a panic for this failure case, when the tuple contains
both non-unifiable types and a null value:

  list(["a"], "b", null) // should not panic

The null value was causing the unification process to result in a list
or set of dynamic type, which causes the conversion functions to pass
through the original value. This meant that in the final conversion
step, we would attempt to construct a list or set of different values,
which panics.

* Update CHANGELOG.md

* cty: Value.HasWhollyKnownType

This tests whether a value contains any unknown values of unknown type.

This is different than just testing if any of the nested types
are DynamicPseudoType, because a null value of
DynamicPseudoType has a different meaning than an
unknown value of DynamicPseudoType: the null value's
type can't become any more "known".

* Update CHANGELOG.md

* v1.5.0

* Update CHANGELOG.md

---------

Co-authored-by: Martin Atkins <mart@degeneration.co.uk>
Co-authored-by: Alisdair McDiarmid <alisdair@users.noreply.github.com>
Co-authored-by: James Bardin <j.bardin@gmail.com>