Forward to v1.5.1 (#14)
* Prepare for a later v1.5.1 release

* function/stdlib: Allow empty maps in merge function

* Update CHANGELOG.md

* function/stdlib: Fix panic with unknown set values

The fix for set unification recently introduced a bug with unknown set
values which would cause a panic. This was triggered by calling
LengthInt on an unknown collection.

This commit fixes this by first verifying that the argument is known, so
that calling its length is safe. If the argument is not known, it could
be of any length, so we do not want to omit its element type from the
set to be unified.

Includes regression tests.

* Update CHANGELOG.md

* v1.5.1

* Update CHANGELOG.md

* go mod tidy

---------

Co-authored-by: Martin Atkins <mart@degeneration.co.uk>
Co-authored-by: Pam Selle <pamela.selle@gmail.com>
Co-authored-by: Alisdair McDiarmid <alisdair@users.noreply.github.com>