Forward to 1.7.0 (#17)
* Prepare for a potential future 1.6.1 release

* build: remove go 1.11 and 1.12 from travis config

* cty: RawEquals correct behavior for sets with partially-unknown values

Sets with the same type and same unknown values are equal enough for
RawEqual's purposes. This commit takes advantage of the setRules's
defined ordering to compare each element of the sets as lists.

* v1.6.1

* Prepare for an assumed future v1.6.2 release

* cty: Utilities to unmark with marks at specific paths

This allows a caller to call a method like Unmark, but instead of getting a superset of marks, get an array of marks and their associated paths. This allows for a Value to be unmarked and then remarked later without losing so much detail about the marks.

* Update CHANGELOG.md

* cty: Fix deep marks functions with new transformer

The UnmarkDeep, UnmarkDeepWithPaths, and MarkWithPaths functions could
previously panic with nested complex values. This was due to the
Transform callback function only being called as part of a postorder
traversal, when it is not permitted to traverse a marked value.

This commit introduces a new TransformWithTransformer function,
requiring an implementation of a new Transformer interface. Using this
interface, callers can implement either preorder or postorder
traversals.

In turn, this allows us to write deep transformation for marks which
successfully cope with all nested structures.

The commit includes significantly more tests for these functions, which
should now cover all complex value types.

* cty: Remove already-fixed FIXME for SetVal

The immediately preceding code already unmarks the values and applies
the marking to the set, so this comment and panic are out of date.

* functions/stdlib: formatdate to correctly handle literals at the end of the format string

* Update CHANGELOG.md

* cty: Fix path array reuse with UnmarkDeepWithPaths

When unmarking a complex value and retaining the marked paths, a
sufficiently deep structure would result in incorrect duplicate path
output. For example, this structure (in HCL syntax):

{
  environment = [
    {
      variables = {
        "x" = 1
        "y" = 2
      }
    }
  ]
}

If the 1 and 2 values are marked, the resulting path value marks from
UnmarkDeepWithPaths would have two entries, as expected. However, both
would have the same Path attribute, like so:

[
  { environment[0].variables["x"], "mark" },
  { environment[0].variables["x"], "mark" },
]

This is caused by calling `append` in the walk transform function with
the same path object repeatedly, which eventually does not result in
array reallocation, and therefore the path object is modified in-place.

* stdlib: Add tests for join function

* function: Deeply unmark function arguments

For functions which do not explicitly support marks, we now deeply
unmark arguments before calling, and reapply those marks to the return
value. This ensures that these functions do not panic with arguments
which are collection values with marked descendants.

This does result in overly conservatively applying marks in some cases,
but that can be fixed with later work to add explicit support for marks
to those functions.

* Update CHANGELOG.md

* v1.7.0

* Update CHANGELOG.md

---------

Co-authored-by: Martin Atkins <mart@degeneration.co.uk>
Co-authored-by: Kristin Laemmert <mildwonkey@users.noreply.github.com>
Co-authored-by: Pam Selle <pamela.selle@gmail.com>
Co-authored-by: Alisdair McDiarmid <alisdair@users.noreply.github.com>
Co-authored-by: OwenTuz <owentuz@gmail.com>