Forward to 1.8.0 (#18)
* Prepare for a potential future 1.6.1 release

* build: remove go 1.11 and 1.12 from travis config

* cty: RawEquals correct behavior for sets with partially-unknown values

Sets with the same type and same unknown values are equal enough for
RawEqual's purposes. This commit takes advantage of the setRules's
defined ordering to compare each element of the sets as lists.

* v1.6.1

* Prepare for an assumed future v1.6.2 release

* cty: Utilities to unmark with marks at specific paths

This allows a caller to call a method like Unmark, but instead of getting a superset of marks, get an array of marks and their associated paths. This allows for a Value to be unmarked and then remarked later without losing so much detail about the marks.

* Update CHANGELOG.md

* cty: Fix deep marks functions with new transformer

The UnmarkDeep, UnmarkDeepWithPaths, and MarkWithPaths functions could
previously panic with nested complex values. This was due to the
Transform callback function only being called as part of a postorder
traversal, when it is not permitted to traverse a marked value.

This commit introduces a new TransformWithTransformer function,
requiring an implementation of a new Transformer interface. Using this
interface, callers can implement either preorder or postorder
traversals.

In turn, this allows us to write deep transformation for marks which
successfully cope with all nested structures.

The commit includes significantly more tests for these functions, which
should now cover all complex value types.

* cty: Remove already-fixed FIXME for SetVal

The immediately preceding code already unmarks the values and applies
the marking to the set, so this comment and panic are out of date.

* functions/stdlib: formatdate to correctly handle literals at the end of the format string

* Update CHANGELOG.md

* cty: Fix path array reuse with UnmarkDeepWithPaths

When unmarking a complex value and retaining the marked paths, a
sufficiently deep structure would result in incorrect duplicate path
output. For example, this structure (in HCL syntax):

{
  environment = [
    {
      variables = {
        "x" = 1
        "y" = 2
      }
    }
  ]
}

If the 1 and 2 values are marked, the resulting path value marks from
UnmarkDeepWithPaths would have two entries, as expected. However, both
would have the same Path attribute, like so:

[
  { environment[0].variables["x"], "mark" },
  { environment[0].variables["x"], "mark" },
]

This is caused by calling `append` in the walk transform function with
the same path object repeatedly, which eventually does not result in
array reallocation, and therefore the path object is modified in-place.

* stdlib: Add tests for join function

* function: Deeply unmark function arguments

For functions which do not explicitly support marks, we now deeply
unmark arguments before calling, and reapply those marks to the return
value. This ensures that these functions do not panic with arguments
which are collection values with marked descendants.

This does result in overly conservatively applying marks in some cases,
but that can be fixed with later work to add explicit support for marks
to those functions.

* Update CHANGELOG.md

* v1.7.0

* Update CHANGELOG.md

* CHANGELOG: Prepare for a potential future v1.7.1 release

* cty: NilType.Equals(...) without a panic

* cty: Maintain precision correctly in Multiply and Modulo

When working with big.Float values, it is important to ensure the
desired precision for the operations is maintained throughout. The
default precision for a Float is set initially by the max argument
precision when creating the non-zero value. If this is coming from a
Float which was set via SetInt, and that Int value was created by
SetInt64, then the precision is only 64 bits.

* Update CHANGELOG.md

* convert: Fix MismatchMessage for object attributes

When determining which object attributes cause a type mismatch, we must
disregard equal attribute types. Previously we would attempt to find a
conversion between them, deciding that the attribute type was a mismatch
if no conversion exists. However, this is invalid if the two types are
equal, because GetConversion returns nil for equal primitive types.

Because of non-deterministic map iteration order, the test case added in
this commit failed approximately 2/3 of the time before the change. It
now passes consistently.

* Update CHANGELOG.md

* functions/stdlib: merge to return empty object if all args are null

* Update CHANGELOG.md

* function/stdlib: coalescelist should not panic when an argument is null

* Update CHANGELOG.md

* v1.7.1 release

* Prepare for a possible forthcoming v1.7.2 release

* function/stdlib: Fix an incorrect test for the

* cty: fix GoString for object types with optional attributes

* Update CHANGELOG.md

* build: Run Travis-CI tests on Go 1.15 and Go 1.16

* v1.7.2

* Unicode 13 support

This includes both using the Unicode 13 normalization rules when we're
running on Go 1.16 or later, and also using the Unicode 13 character
segmentation rules for the stdlib string functions regardless of which
Go version we have selected.

* Update CHANGELOG.md

* v1.8.0

* Update github.com/apparentlymart/go-textseg

---------

Co-authored-by: Martin Atkins <mart@degeneration.co.uk>
Co-authored-by: Kristin Laemmert <mildwonkey@users.noreply.github.com>
Co-authored-by: Pam Selle <pamela.selle@gmail.com>
Co-authored-by: Alisdair McDiarmid <alisdair@users.noreply.github.com>
Co-authored-by: OwenTuz <owentuz@gmail.com>
Co-authored-by: Radek Simko <radek.simko@gmail.com>
Co-authored-by: James Bardin <j.bardin@gmail.com>