cty: Use DeepValues instead of Walk for deep marks inspections
An earlier commit added cty.DeepValues as an iter.Seq2-based alternative
to Walk, and that's much easier to use for deep inspections because it
doesn't require weird trickery to get the Walk to stop at the right time
and smuggle result values out.

However, in the process of doing this I discovered that cty.DeepValues
itself was falling into that very trap of failing to terminate Walk
correctly, and so this also includes a fixed version of that which will
stop yielding new values once commanded to do so.