cty: Use WrangleMarksDeep for UnmarkDeep and UnmarkDeepWithPaths
These functions were previously using the "Transformer" mechanism, which is
very general but as a result ends up always conservatively rebuilding
the entire data structure it's given.

The recently-added WrangleMarksDeep relies on the fact that it can only
possibly change marks (not the actual values) to avoid building a new
value at all in the common case where there are no marks, to rebuild
only the parts of the structure that have mark changes when changes are
being made, and to reuse the internal "payload" values from the source
values while just wrapping them in a new outer cty.Value.

Testing with a benchmark in another codebase that makes heavy use of both
UnmarkDeep and UnmarkDeepWithPaths shows this being a material performance
improvement.

This also includes a bonus fast path for MarkWithPaths where it will skip
doing anything at all if the given slice of ParkValueMarks is empty.